re-work selectors
