+++ web/js/modules/product-picker/formatConfig.js
